+++ viewer/src/rendering/PAGRenderThread.cpp
@@ -28,10 +28,11 @@ PAGRenderThread::PAGRenderThread(PAGView* pagView) : pagView(pagView) {
 auto PAGRenderThread::flush() -> void {
   pagView->pagPlayer->flush();
   double progress = pagView->pagFile->getProgress();
-  int currentFrame = static_cast<int>(std::round((pagView->getTotalFrame() - 1) * progress));
-  int renderingTime = static_cast<int>(pagView->pagPlayer->renderingTime());
-  int presentingTime = static_cast<int>(pagView->pagPlayer->presentingTime());
-  int imageDecodingTime = static_cast<int>(pagView->pagPlayer->imageDecodingTime());
+  int64_t currentFrame =
+      static_cast<int64_t>(std::round((pagView->getTotalFrame().toDouble() - 1) * progress));
+  int64_t renderingTime = pagView->pagPlayer->renderingTime();
+  int64_t presentingTime = pagView->pagPlayer->presentingTime();
+  int64_t imageDecodingTime = pagView->pagPlayer->imageDecodingTime();
   Q_EMIT frameTimeMetricsReady(currentFrame, renderingTime, presentingTime, imageDecodingTime);
   QMetaObject::invokeMethod(pagView, "update", Qt::QueuedConnection);
 }
